+++ rust/catalyst-voting/src/lib.rs
@@ -1,6 +1,5 @@
 //! Voting primitives which are used among Catalyst ecosystem.
 
 pub mod crypto;
-pub mod txs;
 mod utils;
 pub mod vote_protocol;
