--- rust/catalyst-voting/src/txs/mod.rs
@@ -1,3 +1,0 @@
-//! A catalyst transaction objects implementation
-
-pub mod v1;
